requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
SQLAlchemy==2.0.39
asyncpg==0.29.0
python-dotenv==1.0.1
alembic==1.15.1
bcrypt==4.0.1
passlib==1.7.4 
pydantic==2.10.6
pydantic[email]
python-jose[cryptography]
aiohttp-jinja2==1.5.1
jinja2==3.1.3
python-multipart==0.0.6